テスト用のブールマスク生成関数をhelpers.pyに移動し、テストファイルでインポートするように変更
